Refactor: js폴더 리팩토링 - api, components, utils 폴더로 분리
