--- src/main/resources/static/js/employeeUtils.js
@@ -1,5 +1,0 @@
-// 유효성 검사
-export function validateEmployee(employee) {
-	if (!employee.name) return false
-	return true
-}
