+++ src/main/resources/templates/employee/index.html
@@ -125,7 +125,7 @@
                 </div>
                 <div class="btn-container">
                     <button type="submit">등록</button>
-                    <button type="button" onclick="closeModal()">취소</button>
+                    <button type="button">취소</button>
                 </div>
             </form>
         </div>
